Add files via upload
added additional files. 
archive.py is a file for code that is no longer implemented in the main script but I may need to reference to for some reason.
test.py is a space to test the dataframes and make sure everything is there.
urls.txt contains the urls from which I scraped the data. Note that exhibitions and parades have been left out.